TIP-221: Move MetricFactory
